wip: android: Unable to access KeyChain immediately after system boot
There is a massive issue with certificate authentication.  When the system
starts and the Always-on VPN is initiated the key store is locked and
retrieving the certificate chain or private key fails with "KeyChainException:
IllegalStateException: keystore is LOCKED" until the user unlocks the
screen once.

The built-in client will also fail in this situation (e.g. with XAuth
RSA), it tries three times then stops and shows an error notification.

WIP: I guess we could loop around, but should probably install a TUN
device until the connection is up (what IPs? IPv4/IPv6? DNS?).