wip: android: Create temporary TUN device to block traffic when establishing an Always-on VPN
Not sure if the DNS proxy is actually required on Android 7, at least ping
is able to resolve host names even when the filter TUN device is installed.

wip: While we don't stop retrying if the peer is not immediately reachable,
there are some situations where we currently abort the connection (still
with the blocking TUN device no traffic leaves the device until the
error is acknowledged).  For instance, if the peer's address can't be
resolved (could probably be handled via retry_initiate_interval), in any
authentication error (that includes e.g. the peer restarting and not
finding a config yet and returning NO_PROPOSAL_CHOSEN).  Also any errors
like proposal mismatches (IKE or CHILD) are currently reasons to abort.
Changing that would require workarounds in android_service (like
restarting the initiation if these errors occur if always-on is enabled)
or a new option (e.g. on ike_cfg) that would cause the daemon to retry
(according to keyingtries) if such errors occur.  I guess the former is
easier to implement.

Also, currently always-on is only enabled if actually triggered by that
feature. Even when configured as always-on profile manually starting a
profile does not enable these features. I guess we could check if the
started profile is set as always-on profile and enable the features.